+++ src/rule/one_of_rule.cpp
@@ -35,7 +35,7 @@ bool OneOfRule::validate(const Variant &target, ValidationContext &context) cons
 
 	if (passed_count == 0) {
 		context.add_error(
-				vformat("Value failed all %d oneOf schemas", sub_rules.size()),
+				vformat("Value failed all %d oneOf schemas", (int)sub_rules.size()),
 				"oneOf");
 
 		// Show errors from first failed schema for debugging
